structure: minor changes
